Extract V1 setup to new class.
